+++ workspaces/azure-devops/plugins/scaffolder-backend-module-azure-devops/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-scaffolder-backend-module-azure-devops
 
+## 0.6.0
+
+### Minor Changes
+
+- ff23f2f: Backstage version bump to v1.36.1
+
 ## 0.5.0
 
 ### Minor Changes