+++ workspaces/bazaar/plugins/bazaar-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-bazaar-backend
 
+## 0.7.0
+
+### Minor Changes
+
+- bcfde77: Backstage version bump to v1.36.1
+
 ## 0.6.0
 
 ### Minor Changes